sundry build cleanup
Summary:
- reset go-nmap to upstream version now that fix was integrated
- add lint-go target to preserve lint cleanliness
- add lint-all-go target to lint even unclean things
- add base-build.Jenkinsfile jenkins pipeline DSL file, and linter
- consistency improvements in Makefile
- lint clean for some low hanging fruit
- silence annoying warning during pkg construction
- reorganize build/ subdirectory

Test Plan: Build.  Let harbormaster build.  Run smoketest on pi.

Reviewers: stephen, nils, jeff

Reviewed By: stephen, nils, jeff

Subscribers: buildbot

Differential Revision: https://ph0.b10e.net/D122

+++ build/deb-pkg/deb-pkg.py
@@ -12,7 +12,7 @@
 
 """build Debian packages from product software"""
 
-import getopt
+import argparse
 import logging
 import os
 import shutil
@@ -24,7 +24,7 @@ import sh
 logging.basicConfig(level=logging.DEBUG,
                     format="%(asctime)s %(levelname)s %(message)s")
 
-template = """
+TEMPLATE = """
 Package: %s
 Version: %s
 Section: non-free/embedded
@@ -62,7 +62,8 @@ class DebPackage:
     def rm_work_dir(self, arch):
         """Delete the package's work directory."""
         try:
-            shutil.rmtree(self.work_dir(arch))
+            if os.path.exists(self.work_dir(arch)):
+                shutil.rmtree(self.work_dir(arch))
         except Exception as e:
             logging.warning("rmtree %s: %s", self.work_dir(arch), e)
 
@@ -82,7 +83,7 @@ class DebPackage:
         depends = ",".join(self.depends)
 
         controlf = open(self.work_dir(arch) + "/DEBIAN/control", "w")
-        print(template % (self.name, self.version, arch, depends,
+        print(TEMPLATE % (self.name, self.version, arch, depends,
                           "Brightgate Software <contact_us@brightgate.com>",
                           self.description),
               file=controlf)
@@ -90,7 +91,7 @@ class DebPackage:
 
         # copy files in build matching name-* to /DEBIAN/*
         for f in ["prerm", "postinst"]:
-            src = "build/%s-%s" % (self.name, f)
+            src = "build/deb-pkg/%s-%s" % (self.name, f)
             if os.path.exists(src):
                 dst = self.work_dir(arch) + "/DEBIAN/" + f
                 shutil.copyfile(src, dst)
@@ -122,46 +123,39 @@ calver = time.strftime("%y%m%d%H%M")
 
 packages = [
     DebPackage("bg-cloud", "0.0.%s-1" % calver, "amd64", "proto.%s/cloud",
-        ["libc6"],
-        """Cloud components."""),
+               ["libc6"],
+               """Cloud components."""),
     DebPackage("bg-appliance", "0.0.%s-1" % calver, ["armhf", "amd64"],
-        "proto.%s/appliance",
-        ["bridge-utils", "hostapd", "libc6", "libzmq3-dev", "libpcap-dev"],
-        """Appliance components.""")
+               "proto.%s/appliance",
+               ["bridge-utils", "hostapd", "libc6", "libzmq3-dev", "libpcap-dev"],
+               """Appliance components.""")
     ]
 
-if __name__ == "__main__":
-    do_lint = False
-    compresstype = "gzip"
-    compresslevel = 5
+def main_func():
+    """Main program logic"""
+    parser = argparse.ArgumentParser()
+    parser.add_argument('--arch', '-a', required=True)
+    parser.add_argument('--lint', action='store_true')
+    parser.add_argument('--compresslevel', '-z', type=int, default=5)
+    parser.add_argument('--compresstype', '-Z', default="gzip",
+                        choices=["none", "gzip", "xz"])
 
-    opts, pargs = getopt.getopt(sys.argv[1:], "a:Z:z:",
-            longopts=["arch=", "compresstype=", "compresslevel=", "lint"])
-
-    for opt, arg in opts:
-        if opt == "-a" or opt == "--arch":
-            arch = arg
-        elif opt == "--lint":
-            do_lint = True
-        elif opt == "-z" or opt == "--compresslevel":
-            compresslevel = arg
-        elif opt == "-Z" or opt == "--compresstype":
-            if arg not in ["none", "gzip", "xz"]:
-                logging.error("unrecognized compression type '%s'", arg)
-                sys.exit(2)
-            compresstype = arg
+    opts = parser.parse_args(sys.argv[1:])
 
     for p in packages:
-        if arch not in p.arches:
-            logging.info("skipping %s for %s (supports %s)", arch, p.name,
+        if opts.arch not in p.arches:
+            logging.info("skipping %s for %s (supports %s)", opts.arch, p.name,
                          p.arches)
             continue
 
         logging.info("begin %s package build", p.name)
-        p.collect_contents(arch=arch)
-        p.build_package(arch=arch, compresstype=compresstype,
-                        compresslevel=compresslevel)
+        p.collect_contents(arch=opts.arch)
+        p.build_package(arch=opts.arch, compresstype=opts.compresstype,
+                        compresslevel=opts.compresslevel)
         logging.info("end %s package build", p.name)
 
-        if do_lint:
-            p.run_lint(arch)
+        if opts.lint:
+            p.run_lint(opts.arch)
+
+if __name__ == "__main__":
+    main_func()
